+++ presentation/src/commonTest/kotlin/dev/fobo66/valiutchik/presentation/PreferencesViewModelTest.kt
@@ -14,15 +14,13 @@
  *    limitations under the License.
  */
 
-package fobo66.exchangecourcesbelarus.ui.preferences
+package dev.fobo66.valiutchik.presentation
 
 import app.cash.turbine.test
 import dev.fobo66.domain.testing.fake.FakeLoadDefaultCityPreference
 import dev.fobo66.domain.testing.fake.FakeLoadUpdateIntervalPreference
 import dev.fobo66.domain.testing.fake.FakeUpdateDefaultCityPreference
 import dev.fobo66.domain.testing.fake.FakeUpdateUpdateIntervalPreference
-import dev.fobo66.valiutchik.presentation.PreferencesViewModel
-import dev.fobo66.valiutchik.presentation.PreferencesViewModelImpl
 import fobo66.valiutchik.core.KEY_DEFAULT_CITY
 import fobo66.valiutchik.core.KEY_UPDATE_INTERVAL
 import kotlin.test.AfterTest
